Creating the function setStaticIP
I am currently working on a project and needed a function to change the network configuration while the Tiva was active and because of local requeriments we can't use a DHCP server, so I checked the lwiplib library and found the lwIPNetworkConfigChange function and decide to add that functionality to my local Ethernet library. I think this could be a useful function, I saw some people on the internet asking for such functionality.